+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2 (30.12.2011)
+
+* allow static and dynamic routing keys for amqp
+* allow removing middleware
+
 ## 0.0.1 (30.12.2011)
 
 * first extracted version